+++ lib.esm/providers/provider-jsonrpc.js
@@ -635,12 +635,18 @@ export class JsonRpcApiProvider extends AbstractProvider {
                 });
             }
         }
-        if (message.match(/the method .* does not exist/i)) {
+        let unsupported = !!message.match(/the method .* does not exist/i);
+        if (!unsupported) {
+            if (error && error.details && error.details.startsWith("Unauthorized method:")) {
+                unsupported = true;
+            }
+        }
+        if (unsupported) {
             return makeError("unsupported operation", "UNSUPPORTED_OPERATION", {
-                operation: payload.method, info: { error }
+                operation: payload.method, info: { error, payload }
             });
         }
-        return makeError("could not coalesce error", "UNKNOWN_ERROR", { error });
+        return makeError("could not coalesce error", "UNKNOWN_ERROR", { error, payload });
     }
     /**
      *  Requests the %%method%% with %%params%% via the JSON-RPC protocol
@@ -809,7 +815,7 @@ function spelunkData(value) {
         return null;
     }
     // These *are* the droids we're looking for.
-    if (typeof (value.message) === "string" && value.message.match("reverted") && isHexString(value.data)) {
+    if (typeof (value.message) === "string" && value.message.match(/revert/i) && isHexString(value.data)) {
         return { message: value.message, data: value.data };
     }
     // Spelunk further...